Move a bunch of test stuff into a base class.
